+++ singletons/global.gd
@@ -83,13 +83,6 @@ signal close_admin_popups_required()
 signal show_hide_gui_requested(is_toggle: bool, is_show: bool)
 
 
-## Sizes available for setting "gui_size". See also [member IVCoreSettings.gui_size_multipliers].
-enum GUISize {
-	GUI_SMALL,
-	GUI_MEDIUM,
-	GUI_LARGE,
-}
-
 ## Resolution choice for the background starmap; see [IVAssetPreloader].
 enum StarmapSize {
 	STARMAP_8K,