+++ tree/body.gd
@@ -215,7 +215,6 @@ const PERSIST_PROPERTIES: Array[StringName] = [
 static var replacement_subclass: Script
 ## Set this script to replace the [IVPhysicalBody] class.
 static var replacement_physical_body_class: Script
-
 ## Static class setting. Default value is a dashed circle.
 static var default_symbol := "\u25CC"
 ## Static class setting.
@@ -226,16 +225,12 @@ static var max_hud_dist_orbit_radius_multiplier := 100.0
 static var min_hud_dist_radius_multiplier := 500.0
 ## Static class setting.
 static var min_hud_dist_star_multiplier := 20.0
-
-# TEST: It's possible that self-class referencing in static vars below is causing
-# editor leaks on editor exit in 4.5.x. Test if they don't go away in 4.6.
-
 ## A static class dictionary that contains all added IVBody instances.
+## WARNING: Access on main thread only!
 static var bodies: Dictionary[StringName, IVBody] = {}
-
 ## A static class dictionary that contains IVBody instances that are at the top
 ## of a system, i.e., the primary star for every star system (these have no
-## [IVOrbit]).
+## [IVOrbit]). WARNING: Access on main thread only!
 static var top_bodies: Dictionary[StringName, IVBody] = {}
 
 static var _selection_ordered_bodies: Array[IVBody] = [] # build/rebuild only when needed
@@ -326,13 +321,23 @@ var _stroboscope_rotation := 0.0
 # *****************************************************************************
 # create methods
 
-## Creates a new [IVOrbit] instance (or specified [member replacement_subclass])
+## Creates a new [IVBody] instance (or specified [member replacement_subclass])
 ## using specified parameters. See also [method create_from_astronomy_specs].
 @warning_ignore("shadowed_variable", "shadowed_variable_base_class")
-static func create(name: StringName, flags: int, mean_radius: float, gravitational_parameter: float,
-		orientation_at_epoch: Basis, rotation_axis: Vector3, rotation_at_epoch: float,
-		rotation_rate: float, orbit: IVOrbit, characteristics: Dictionary,
-		components: Dictionary, exisiting_body: IVBody = null) -> IVBody:
+static func create(
+		name: StringName,
+		flags: int,
+		mean_radius: float,
+		gravitational_parameter: float,
+		orientation_at_epoch: Basis,
+		rotation_axis: Vector3,
+		rotation_at_epoch: float,
+		rotation_rate: float,
+		orbit: IVOrbit,
+		characteristics: Dictionary,
+		components: Dictionary,
+		exisiting_body: IVBody = null
+	) -> IVBody:
 	
 	assert(name)
 	assert(mean_radius > 0.0, "IVBody requires mean_radius > 0.0")
@@ -375,7 +380,7 @@ static func create(name: StringName, flags: int, mean_radius: float, gravitation
 	return body
 
 
-## Creates a new [IVOrbit] instance (or specified [member replacement_subclass])
+## Creates a new [IVBody] instance (or specified [member replacement_subclass])
 ## using specified parameters. [param right_ascension] and [param declination]
 ## define "North" for this body. If [param rotation_period] is negative, then
 ## this body has retrograde rotation (e.g., Venus). If [param flags] & BODYFLAGS_TIDALLY_LOCKED,